autotools build of bitcoind
just a start, needs more detections at configure, however works to build the daemon alone for now.
more flags should be present on ./configure to select gui build and detect required libraries.
libtool is just used to build the internal libcryptopp, but we can ship libbitcoin.so in future.

the aim of this fork is mostly that of GNUification for bitcoin codebase and process separation.
code will be shared between daemon and client implementations via shared libraries.

it might look as a traumatic for now, but well think of the long term...
win32 and osx builds can be fixed to work using autotools or a parallel cmake build.