force use of a specific version of libdb: 4.8
the choice of 4.8 is made on the fact that debian uses that and it is the higher version available. libdb upgrades the database dropping backward compatibility, so we need to stay on the highest option. however, this is a disruptive policy of libdb for wider and smoother deploy, suggesting its urgent substitution.